rope: moved a trait impl for Node into `internals` module
